Fix: Vehicle errors are always JSON objects
It works, however: sometimes they are displayed as Error: error: <error message> due to inconsistencies in how I wrote it and how it displays the errors, such as displaying only the value or the whole error property including the key. The AJV errors have a similar issue. Check entries too.