feat: add NewEntryForm (needs refactoring)
Again, it's copy pasted and insanely long. I need to create a separate component for the entries. Maybe I could even combine all the Forms since they're so similar and simply have them be changed as needed. Hardcoding all this stuff sucks and makes the code hard to maintain, since I have to tweak 4 separate forms to fix ONE issue. But it'll have to do for now.