fix: NewEntryForm and handleCreateEntry date improvements (needs refactoring)
NewEntryForm now displays separate fields for day,month,year, and then handleCreateEntry combines them into a YYYY-MM-DD string. However, it doesn't validate whether the date is actually valid, such as February 31st, or leap years, or months which only have 30 days. There is also a bug where it can take 'e' as an input into the field, and possibly other non-numeric characters. But under normal circumstances it does work, so this is just another addition to the pile of technical debt.