Improved form state system and bugfixes (Needs simplifying)
This new implementation uses like 6 different state variables to track which form should be displayed. This is TERRIBLE, but it could work with a small app. My issue was that I was always treating the states as booleans, leading to a really complicated system where I had to specify which forms to show and hide. Instead, have the state variable store a string, like newVehicle, updateVehicle, etc. which will make this 10 times easier. For example: To add a new vehicle, it'll check if activeForm state is 'newVehicle', then it'll render it. Once the user cancels, submits, or otherwise leaves the form, the activeForm state will have to be set to null again. Great solution!