fix: show notification in top right, minor bug fixes
Notification component now works correctly, but it only displays one notification at a time (the latest one) and overwrites the old ones. This is because it only saves the state of one notification instead of an array. However, this is a very minor issue.

Fixed some minor issues with VehicleEntries useEffect giving warnings. No change in code functionality.

Creating a new entry via handleCreateEntry now adds the new entry to the top of the entry list. This is unsorted, but once the page is refreshed it appears in the correct position. The backend sorts it, but the frontend does not. I don't consider this a bug, since when you make a new entry and suddenly realize you made an error, you can easily edit it again. Besides, the user will usually add entries that are newest.