fix: VehicleEntries table max dimensions
Several pieces of the table including the table itself now have size limits. The notes section will be limited as it's the one with the highest risk of being too large. This should be improved upon such as by adding limits to the other columns and properly calculating how big each part should be (both by default, maximum and minimum), but under normal conditions it's sufficient.